+++ packages/smarthr-ui/src/components/Dialog/RemoteTriggerActionDialog/index.ts
@@ -1,1 +1,0 @@
-export { RemoteTriggerActionDialog } from './RemoteTriggerActionDialog'