+++ packages/smarthr-ui/src/components/Disclosure/index.ts
@@ -1,0 +1,1 @@
+export { DisclosureTrigger, DisclosureContent } from './Disclosure'