+++ packages/smarthr-ui/src/components/Icon/Icon.tsx
@@ -62,6 +62,7 @@ import {
   FaClipboard,
   FaClipboardCheck,
   FaClipboardList,
+  FaClipboardQuestion,
   FaClock,
   FaClockRotateLeft,
   FaCloudArrowDown,
@@ -104,7 +105,6 @@ import {
   FaGear,
   FaGears,
   FaGlobe,
-  FaGoogleDrive,
   FaGraduationCap,
   FaGrip,
   FaGripLines,
@@ -261,6 +261,7 @@ export const FaCircleXmarkIcon = /*#__PURE__*/ generateIcon(FaCircleXmark)
 export const FaClipboardCheckIcon = /*#__PURE__*/ generateIcon(FaClipboardCheck)
 export const FaClipboardIcon = /*#__PURE__*/ generateIcon(FaClipboard)
 export const FaClipboardListIcon = /*#__PURE__*/ generateIcon(FaClipboardList)
+export const FaClipboardQuestionIcon = /*#__PURE__*/ generateIcon(FaClipboardQuestion)
 export const FaClockIcon = /*#__PURE__*/ generateIcon(FaClock)
 export const FaClockRotateLeftIcon = /*#__PURE__*/ generateIcon(FaClockRotateLeft)
 export const FaCloudArrowDownIcon = /*#__PURE__*/ generateIcon(FaCloudArrowDown)
@@ -305,7 +306,6 @@ export const FaGaugeIcon = /*#__PURE__*/ generateIcon(FaGauge)
 export const FaGearIcon = /*#__PURE__*/ generateIcon(FaGear)
 export const FaGearsIcon = /*#__PURE__*/ generateIcon(FaGears)
 export const FaGlobeIcon = /*#__PURE__*/ generateIcon(FaGlobe)
-export const FaGoogleDriveIcon = /*#__PURE__*/ generateIcon(FaGoogleDrive)
 export const FaGraduationCapIcon = /*#__PURE__*/ generateIcon(FaGraduationCap)
 export const FaGripIcon = /*#__PURE__*/ generateIcon(FaGrip)
 export const FaGripLinesIcon = /*#__PURE__*/ generateIcon(FaGripLines)