+++ packages/smarthr-ui/src/themes/__tests__/createBreakpoint.ts
@@ -1,15 +1,6 @@
 import { createBreakpoint } from '../createBreakpoint'
-import { createSize } from '../createSize'
 
 describe('createBreakpoint', () => {
-  it('returns same break point theme with createSize', () => {
-    const actual = createBreakpoint()
-    const expected = createSize()
-
-    expect(actual.SP).toBe(expected.mediaQuery.SP)
-    expect(actual.TABLET).toBe(expected.mediaQuery.TABLET)
-  })
-
   it('returns customized break point theme when give user break point', () => {
     const actual = createBreakpoint({
       SP: 100,